Fix exception in ClusterMemoryManager.process
The result of isLastKilledQueryGone is inverted for common case. As a result,
ClusterMemoryManager.process could run lowMemoryKiller.chooseQueryToKill
before it has updated the nodes and pools fields. This will lead to
NoSuchElementException when ClusterMemoryManager tries to map chosen query
id to the corresponding QueryExecution. This will skip updatePools and
updateNodes, causing ClusterMemoryManager to stop functioning permanently.